refactor: Extract TvShowLayOut to separate composable
